add dm_to_dd to lib tests, fix 2 mistakes in the reading tests
